scraping-api/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.4
debugpy==1.6.7
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.3.2
Flask-Cors==4.0.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter-client==8.3.0
jupyter-core==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.7
openpyxl==3.1.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip-autoremove==0.10.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pypdf2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing-extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2
